# Documentation/api/openapi/weatherwear-api.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WeatherWear API",
    "description": "REST API for user authentication, weather lookup, AI clothing recommendations, recommendation history, feedback, and an AI style chat assistant.",
    "version": "1.0.0",
    "contact": {
      "name": "Dubrovin Mikita",
      "email": "dubrovin.nikita@student.ehu.lt"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8090/api",
      "description": "Local development server"
    },
    {
      "url": "https://wwtodipl-production.up.railway.app/api",
      "description": "Production server configured by application-prod.yml"
    }
  ],
...
  "paths": {
...
    "/weather": {
      "get": {
        "tags": [
          "Weather"
        ],
        "operationId": "getWeatherByCity",
        "summary": "Get weather by city",
        "description": "Returns the current weather for a city. The backend reuses cached weather records for 30 minutes.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "city",
            "in": "query",
            "required": true,
            "description": "City name. Maximum length is 120 characters.",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 120
            },
            "example": "Vilnius"
          }
        ],
        "responses": {
          "200": {
            "description": "Current weather for the requested city.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeatherResponse"
                },
                "examples": {
                  "weather": {
                    "value": {
                      "city": "Vilnius",
                      "latitude": 54.6872,
                      "longitude": 25.2797,
                      "temperature": 12.4,
                      "feelsLike": 10.8,
                      "humidity": 71,
                      "windSpeed": 4.6,
                      "condition": "CLOUDS",
                      "precipitation": 0,
                      "cached": false
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "502": {
            "$ref": "#/components/responses/BadGatewayError"
          }
        }
      }
    },
    "/weather/coordinates": {
      "get": {
        "tags": [
          "Weather"
        ],
        "operationId": "getWeatherByCoordinates",
        "summary": "Get weather by coordinates",
        "description": "Returns current weather for latitude and longitude. Latitude must be between -90 and 90; longitude must be between -180 and 180.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double",
              "minimum": -90,
              "maximum": 90
            },
            "example": 54.6872
          },
          {
            "name": "lon",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double",
              "minimum": -180,
              "maximum": 180
            },
            "example": 25.2797
          }
        ],
        "responses": {
          "200": {
            "description": "Current weather for the requested coordinates.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeatherResponse"
                },
                "examples": {
                  "weatherByCoordinates": {
                    "summary": "Weather by coordinates from cache",
                    "value": {
                      "city": "Vilnius",
                      "latitude": 54.6872,
                      "longitude": 25.2797,
                      "temperature": 12.1,
                      "feelsLike": 10.5,
                      "humidity": 72,
                      "windSpeed": 4.2,
                      "condition": "CLOUDS",
                      "precipitation": 0,
                      "cached": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "502": {
            "$ref": "#/components/responses/BadGatewayError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT returned by /auth/register or /auth/login. Send it as Authorization: Bearer <token>."
      }
    },
    "responses": {
      "ValidationError": {
        "description": "Request validation failed, a required parameter is missing, or the JSON body is malformed.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "validation": {
                "value": {
                  "timestamp": "2026-04-30T12:45:00",
                  "status": 400,
                  "error": "Bad Request",
                  "message": "email: Email must be valid",
                  "path": "/api/auth/register"
                }
              }
            }
          }
        }
      },
      "UnauthorizedError": {
        "description": "Authentication is missing, invalid, or expired.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "unauthorized": {
                "value": {
                  "timestamp": "2026-04-30T12:45:00",
                  "status": 401,
                  "error": "Unauthorized",
                  "message": "Authentication is required",
                  "path": "/api/preferences"
                }
              }
            }
          }
        }
      },
...
      "BadGatewayError": {
        "description": "An upstream dependency such as OpenWeather or the LLM provider failed.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "upstream": {
                "value": {
                  "timestamp": "2026-04-30T12:45:00",
                  "status": 502,
                  "error": "Bad Gateway",
                  "message": "Failed to fetch weather data",
                  "path": "/api/weather"
                }
              }
            }
          }
        }
      },
      "ServerError": {
        "description": "Unexpected server error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "serverError": {
                "value": {
                  "timestamp": "2026-04-30T12:45:00",
                  "status": 500,
                  "error": "Internal Server Error",
                  "message": "Unexpected server error",
                  "path": "/api/recommendations"
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
...
      "WeatherResponse": {
        "type": "object",
        "properties": {
          "city": {
            "type": "string",
            "example": "Vilnius"
          },
          "latitude": {
            "type": "number",
            "format": "double",
            "example": 54.6872
          },
          "longitude": {
            "type": "number",
            "format": "double",
            "example": 25.2797
          },
          "temperature": {
            "type": "number",
            "format": "double",
            "description": "Temperature in Celsius.",
            "example": 12.4
          },
          "feelsLike": {
            "type": "number",
            "format": "double",
            "description": "Feels-like temperature in Celsius.",
            "example": 10.8
          },
          "humidity": {
            "type": "integer",
            "format": "int32",
            "description": "Relative humidity percentage.",
            "example": 71
          },
          "windSpeed": {
            "type": "number",
            "format": "double",
            "description": "Wind speed in meters per second.",
            "example": 4.6
          },
          "condition": {
            "$ref": "#/components/schemas/WeatherCondition"
          },
          "precipitation": {
            "type": "number",
            "format": "double",
            "description": "Precipitation volume from the upstream provider when available.",
            "example": 0
          },
          "cached": {
            "type": "boolean",
            "description": "True when the response came from the local weather cache.",
            "example": false
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "timestamp",
          "status",
          "error",
          "message",
          "path"
        ],
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2026-04-30T12:45:00"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "example": 400
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "message": {
            "type": "string",
            "example": "email: Email must be valid"
          },
          "path": {
            "type": "string",
            "example": "/api/auth/register"
          }
        }
      },
...
      "WeatherCondition": {
        "type": "string",
        "enum": [
          "CLEAR",
          "CLOUDS",
          "RAIN",
          "DRIZZLE",
          "THUNDERSTORM",
          "SNOW",
          "MIST",
          "FOG",
          "WIND",
          "UNKNOWN"
        ],
        "example": "CLOUDS"
      },
...
    }
  }
}